ownline-backend/requirements.txt
Flask
flask-jwt-extended
tinydb
ujson
python-telegram-bot
pycrypto